feat: comments for answers, a basic text input component
chore: close eslint rule react/prop-types
