--- src/component/Editor/index.js
@@ -1,1 +1,0 @@
-export { default as Editor } from './Editor';
