+++ src/component/index.js
@@ -1,0 +1,5 @@
+// Better sort 'export' alphabetically 😀
+export { default as Comments } from './Comment/Comment';
+export { default as CommentButton } from './Button/CommentButton';
+export { default as Editor } from './Editor/Editor';
+export { default as MdEditor } from './Editor/MdEditor';
